requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
boto==2.49.0
certifi==2022.12.7
cffi==1.15.1
cloudinary==1.30.0
confusable-homoglyphs==3.2.0
cryptography==38.0.4
dj-database-url==1.2.0
Django==4.1.4
django-bootstrap3==22.2
django-bootstrap4==22.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-registration==3.3
django-require==1.0.11
django-require-s3==1.0.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-storages==1.13.1
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-api-key==2.2.0
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
importlib-metadata==5.2.0
numpy==1.24.0
opencv-python==4.6.0.66
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==1.7.1
pyperclip==1.8.2
python-decouple==3.1
pytz==2022.7
sh==1.14.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
toml==0.10.2
typing-extensions==4.4.0
urllib3==1.26.13
waitress==2.1.2
whitenoise==6.2.0
zipp==3.11.0
